+++ backend/src/main/java/com/projetCloud/projetCloud/repository/signalement/SignalementHistoriqueRepository.java
@@ -3,5 +3,5 @@ package com.projetCloud.projetCloud.repository.signalement;
 import com.projetCloud.projetCloud.model.signalement.SignalementHistorique;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SignalementHistoriqueRepository extends JpaRepository<SignalementHistorique, Long> {
+public interface SignalementHistoriqueRepository extends JpaRepository<SignalementHistorique, Integer> {
 }
